# openapi.yaml
openapi: 3.0.1
info:
  title: Paperless Rest Server
  version: '1.0'
paths:
  /api/documents:
    get:
      tags:
        - Documents
      operationId: GetDocuments
      parameters:
        - name: Page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: page_size
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: query
          in: query
          style: form
          schema:
            type: string
        - name: ordering
          in: query
          style: form
          schema:
            type: string
        - name: tags__id__all
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
              format: int32
        - name: document_type__id
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: storage_path__id__in
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: correspondent__id
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: truncate_content
          in: query
          style: form
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  next:
                    type: integer
                  previous:
                    type: integer
                  all:
                    type: array
                    items:
                      type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        correspondent:
                          type: integer
                        document_type:
                          type: integer
                        storage_path:
                          type: integer
                        title:
                          type: string
                        content:
                          type: string
                        tags:
                          type: array
                          items:
                            type: integer
                        created:
                          type: string
                        created_date:
                          type: string
                        modified:
                          type: string
                        added:
                          type: string
                        archive_serial_number:
                          type: integer
                        original_file_name:
                          type: string
                        archived_file_name:
                          type: string
                        owner:
                          type: integer
                        user_can_change:
                          type: boolean
                        notes:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                              note:
                                type: string
                              created:
                                type: string
                              document:
                                type: integer
                              user:
                                type: integer
                            required:
                              - id
                              - note
                              - created
                              - document
                              - user
                      required:
                        - id
                        - correspondent
                        - document_type
                        - storage_path
                        - title
                        - content
                        - tags
                        - created
                        - created_date
                        - modified
                        - added
                        - archive_serial_number
                        - original_file_name
                        - archived_file_name
                        - owner
                        - user_can_change
                        - notes
                required:
                  - count
                  - next
                  - previous
                  - all
                  - results
              examples:
                default:
                  value: >-
                    {"count":1,"next":null,"previous":null,"all":[1],"results":[{"id":1,"correspondent":1,"document_type":1,"storage_path":1,"title":"test","content":"test","tags":[1],"created":"2023-09-08T00:00:00Z","created_date":"2023-09-08","modified":"2023-09-08T12:15:24.652098Z","added":"2023-09-08T12:05:28.338799Z","archive_serial_number":1,"original_file_name":"test.pdf","archived_file_name":"2023-09-08
                    edtest
                    test.pdf","owner":3,"user_can_change":true,"notes":[{"id":1,"note":"pdf
                    in
                    testdir","created":"2023-09-08T12:11:05.848131Z","document":1,"user":3},{"id":2,"note":"pdf
                    2 in
                    test","created":"2023-09-08T12:15:12.328015Z","document":1,"user":3}]}]}
...
  /api/documents/post_document:
    post:
      tags:
        - Documents
      operationId: UploadDocument
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                created:
                  type: string
                  format: date-time
                document_type:
                  type: integer
                  format: int32
                tags:
                  type: array
                  items:
                    type: integer
                    format: int32
                correspondent:
                  type: integer
                  format: int32
                document:
                  type: array
                  items:
                    type: string
                    format: binary
            encoding:
              title:
                style: form
              created:
                style: form
              document_type:
                style: form
              tags:
                style: form
              correspondent:
                style: form
              document:
                style: form
      responses:
        '200':
          description: Success
  /api/documents/{id}/metadata:
    get:
      tags:
        - Documents
      operationId: GetDocumentMetadata
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  original_checksum:
                    type: string
                  original_size:
                    type: integer
                  original_mime_type:
                    type: string
                  media_filename:
                    type: string
                  has_archive_version:
                    type: boolean
                  original_metadata:
                    type: array
                    items: {}
                  archive_checksum:
                    type: string
                  archive_media_filename:
                    type: string
                  original_filename:
                    type: string
                  lang:
                    type: string
                  archive_size:
                    type: integer
                  archive_metadata:
                    type: array
                    items:
                      type: object
                      properties:
                        namespace:
                          type: string
                        prefix:
                          type: string
                        key:
                          type: string
                        value:
                          type: string
                      required:
                        - namespace
                        - prefix
                        - key
                        - value
                required:
                  - original_checksum
                  - original_size
                  - original_mime_type
                  - media_filename
                  - has_archive_version
                  - original_metadata
                  - archive_checksum
                  - archive_media_filename
                  - original_filename
                  - lang
                  - archive_size
                  - archive_metadata
              examples:
                default:
                  value: >-
                    {"original_checksum":"39b80972dc7f1e975d9d2e8b92ce4e8b","original_size":11868,"original_mime_type":"application/pdf","media_filename":"testdirp.pdf","has_archive_version":true,"original_metadata":[],"archive_checksum":"3597d60204a4b200b97e728622f704ab","archive_media_filename":"testdirp.pdf","original_filename":"test.pdf","lang":"et","archive_size":15561,"archive_metadata":[{"namespace":"http://ns.adobe.com/pdf/1.3/","prefix":"pdf","key":"Producer","value":"pikepdf
                    7.2.0"},{"namespace":"http://ns.adobe.com/xap/1.0/","prefix":"xmp","key":"ModifyDate","value":"2023-09-08T12:04:59+00:00"},{"namespace":"http://ns.adobe.com/xap/1.0/","prefix":"xmp","key":"CreateDate","value":"2023-09-08T11:46:26+00:00"},{"namespace":"http://ns.adobe.com/xap/1.0/","prefix":"xmp","key":"CreatorTool","value":"ocrmypdf
                    14.3.0 / Tesseract OCR-PDF
                    5.3.0"},{"namespace":"http://ns.adobe.com/xap/1.0/mm/","prefix":"xmpMM","key":"DocumentID","value":"uuid:7af613b2-865c-11f9-0000-925a0e2aa608"},{"namespace":"http://purl.org/dc/elements/1.1/","prefix":"dc","key":"format","value":"application/pdf"},{"namespace":"http://purl.org/dc/elements/1.1/","prefix":"dc","key":"title","value":"Ohne
                    Titel"},{"namespace":"http://www.aiim.org/pdfa/ns/id/","prefix":"pdfaid","key":"part","value":"2"},{"namespace":"http://www.aiim.org/pdfa/ns/id/","prefix":"pdfaid","key":"conformance","value":"B"},{"namespace":"http://purl.org/dc/elements/1.1/","prefix":"dc","key":"creator","value":"None"},{"namespace":"http://ns.adobe.com/xap/1.0/","prefix":"xmp","key":"MetadataDate","value":"2023-09-08T12:04:59.067651+00:00"}]}
...